:twisted_rightwards_arrows: Refactor of MedicinePage and OtcPage
